requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0

